+++ theming/theme-classic-tailwind/src/ui-default-options-event-calendar.ts
@@ -50,7 +50,10 @@ export const params: EventCalendarOptionParams = {
   eventColor: 'var(--fc-classic-event)',
   eventContrastColor: 'var(--fc-classic-event-contrast)',
   bgEventColor: 'var(--fc-classic-background-event)',
-  bgEventBgClass: 'bg-[color-mix(in_oklab,var(--fc-event-color)_15%,transparent)]',
+  bgEventBgClass: 'bg-[color-mix(in_oklab,var(--fc-event-color)_var(--fc-classic-background-event-opacity),transparent)]',
+  bgEventFgOpacityClass: 'opacity-(--fc-classic-background-event-foreground-opacity)',
+  smallDotBorderClass: 'border-[calc(var(--fc-classic-small-dot-width)/2)]',
+  largeDotBorderClass: 'border-[calc(var(--fc-classic-large-dot-width)/2)]',
 
   // misc content
   highlightClass: 'bg-(--fc-classic-highlight)',